Use filter instead of where (#620)
Use a plain filter-step instead of where when the predicate is a traversal. The reason is that where has some implications when e.g. the first step in the child traversal of where is an as step. See https://groups.google.com/g/gremlin-users/c/Ju3iSIVohaY for the rationale.